deployment/requirements_hf.txt
# HuggingFace Spaces Requirements
# Minimal requirements for the demo app

streamlit==1.28.0
torch>=2.1.0
torchvision>=0.16.0
timm==1.0.22
transformers==4.57.3
Pillow==10.4.0
numpy==1.24.3
requests==2.31.0
grad-cam>=1.4.0
opencv-python-headless>=4.8.0
